Gets everything working for js
This is the initial implementation of the whole system
for js. It reads config, it pushes changes... it makes
the things.

It is a really MVP implementation and will need work to
factor out things and clean them up for broader use. But
for gerrit shops it should work really well.